+++ docs/conf.py
@@ -10,6 +10,7 @@ extensions = [
     "sphinx_copybutton",
     "sphinx.ext.intersphinx",
     "sphinx.ext.autosectionlabel",
+    "sphinx.ext.napoleon",
 ]
 source_suffix = ".rst"
 master_doc = "index"