+++ src/icalendar/__init__.py
@@ -15,6 +15,7 @@ from icalendar.cal import (
     TimezoneStandard,
     Todo,
 )
+from icalendar.enums import CUTYPE, FBTYPE, PARTSTAT, RANGE, RELATED, RELTYPE, ROLE
 from icalendar.error import (
     ComponentEndMissing,
     ComponentStartMissing,
@@ -111,5 +112,13 @@ __all__ = [
     "ComponentStartMissing",
     "IncompleteAlarmInformation",
     "LocalTimezoneMissing",
+    "CUTYPE",
+    "FBTYPE",
+    "PARTSTAT",
+    "RANGE",
     "vSkip",
+    "RELATED",
+    "vSkip",
+    "RELTYPE",
+    "ROLE",
 ]